+++ example/set_visibility_time_out.dart
@@ -11,20 +11,25 @@ void main() async {
       port: 5460);
   print("Start ...");
   try {
+    // Create a pgmq connexion
     final pgmq = await Pgmq.createConnection(param: databaseParam);
 
-    final queue = await pgmq.createQueue(queueName: 'yaya');
+    // Create a queue
+    final queue = await pgmq.createQueue(queueName: 'queueName');
+
     // Purge queue
     await queue.purgeQueue();
 
+    // Play with pausablePull
     final (pausableTimer, stream) = queue.pausablePull(
         duration: Duration(milliseconds: 100),
         visibilityDuration: Duration(seconds: 30));
     pausableTimer.start();
 
+    // Listen to stream then set vt
     stream.listen((event) async {
       if (event.payload['id'] is int && (event.payload['id'] % 2) == 0) {
-        await queue.setVisibilityTimout(
+        await queue.setVisibilityTimeout(
             messageID: event.messageID, duration: Duration(seconds: 10));
       }
       final dt = DateTime.now();
@@ -38,12 +43,6 @@ void main() async {
       final payload = {'id': i, 'message': 'message $i'};
       await queue.send(payload);
     }
-
-    // final data = (await queue.read(maxReadNumber: 5));
-
-    // for (final msg in data ?? <Message>[]) {
-    //   print(msg.payload);
-    // }
   } catch (e, stackTrace) {
     print(stackTrace);
     print(e.toString());