+++ CHANGELOG.md
@@ -5,4 +5,4 @@ Initial version.
 - feat: Includes the ability to create a `PgmqConnexion`.
 - feat: Includes the ability to create, purge, drop a `Queue`.
 - feat: Includes the ability to send, read, pop, archive, delete a `Message`.
-- feat: Includes the ability to set a `vtc`.
+- feat: Includes the ability to set a `vt` (visibility timeout).